PPM moved into disease model
I've moved the ppm coefficient into the disease model. The reason is that the logic was previously bad. If a = ppm_coeff then our probability of infection was a(1-(1-p)^c) whereas it should be (1-(1-a*p)^c). These are only equal if c (the number of contagious in a given location) is 1. Generally this need not be the case.